requirements.txt
python-telegram-bot[all]>=22.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
